extracted new rocketchat api objects; added support for emojis; code cleanup
